more pointless testing

+++ b_soundbytes/src/main.rs
@@ -1,7 +1,7 @@
 
 
 //Prints it out as a graph.
-fn update_display(byte_stream:&[i16;64])
+fn update_display(byte_stream:&[i8;64])
 {
     for i in (0..13).rev() {
         for j in 0..63 {
@@ -26,14 +26,14 @@ fn compressor_node(threshold:i16, ratio:i16, mut &_byte_stream:[i16;64])
 }
 */
 
-fn blip(position:usize, magnitude:i16, byte_stream:&mut [i16;64])
+fn blip(position:usize, magnitude:i8, byte_stream:&mut [i8;64])
 {
     byte_stream[position] = magnitude;
 
     //Calculate Falloff
     for i in 0..5{
-        byte_stream[(position + i)] = magnitude/i16::from(i);
-        byte_stream[(position - i)] = magnitude/i16::from(i);
+        byte_stream[(position + i)] = magnitude/i8::from(i);
+        byte_stream[(position - i)] = magnitude/i8::from(i);
     }
 }
 
@@ -41,7 +41,7 @@ fn blip(position:usize, magnitude:i16, byte_stream:&mut [i16;64])
 fn main() {
     
     //Test cases can change here.
-    let mut stream:[i16;64] = [0;64];
+    let mut stream:[i8;64] = [0;64];
 
     blip(22, 5, &mut stream);
